Split tunnel from webserver file
