+++ pkg/client/webserver.go
@@ -7,17 +7,13 @@ import (
 	"fmt"
 	"net"
 	"net/http"
-	"net/url"
 	"os"
 	"path"
-	"strings"
 	"time"
 
 	"github.com/Notifiarr/notifiarr/pkg/mnd"
-	"github.com/Notifiarr/notifiarr/pkg/website/clientinfo"
 	"github.com/gorilla/mux"
 	apachelog "github.com/lestrrat-go/apache-logformat/v2"
-	mulery "golift.io/mulery/client"
 )
 
 // StartWebServer starts the web server.
@@ -62,90 +58,6 @@ func (c *Client) StartWebServer(ctx context.Context) {
 	go c.runWebServer()
 }
 
-func (c *Client) startTunnel(ctx context.Context) {
-	// If clientinfo is nil, then we probably have a bad API key.
-	ci := clientinfo.Get()
-	if ci == nil {
-		c.Errorf("Skipping tunnel creation because there is no client info.")
-		return
-	}
-
-	const (
-		maxPoolSize = 25 // maximum websocket connections to the origin (mulery server).
-		maxPoolMin  = 4  // maximum is calculated, and this is the minimum it may be.
-	)
-
-	poolmax := len(c.Config.Apps.Sonarr) + len(c.Config.Apps.Radarr) + len(c.Config.Apps.Lidarr) +
-		len(c.Config.Apps.Readarr) + len(c.Config.Apps.Prowlarr) + len(c.Config.Apps.Deluge) +
-		len(c.Config.Apps.Qbit) + len(c.Config.Apps.Rtorrent) + len(c.Config.Apps.SabNZB) +
-		len(c.Config.Apps.NZBGet) + 1
-
-	if c.Config.Apps.Plex.Enabled() {
-		poolmax++
-	}
-
-	if c.Config.Apps.Tautulli.Enabled() {
-		poolmax++
-	}
-
-	if poolmax > maxPoolSize || ci.IsSub() {
-		poolmax = maxPoolSize
-	} else if poolmax < maxPoolMin {
-		poolmax = maxPoolMin
-	}
-
-	hostname, _ := os.Hostname()
-	if hostInfo, err := c.clientinfo.GetHostInfo(ctx); err != nil {
-		hostname = hostInfo.Hostname
-	}
-
-	// This apache logger is only used for client->server websocket-tunneled requests.
-	remWs, _ := apachelog.New(`%{X-Forwarded-For}i %{X-User-ID}i env:%{X-User-Environment}i %t "%r" %>s %b ` +
-		`"%{X-Client-ID}i" "%{User-agent}i" %{X-Request-Time}i %{ms}Tms`)
-	//nolint:gomnd // just attempting a tiny bit of splay.
-	c.tunnel = mulery.NewClient(&mulery.Config{
-		Name:          hostname,
-		ID:            c.Config.HostID,
-		ClientIDs:     []any{ci.User.ID},
-		Targets:       ci.User.Tunnels,
-		PoolIdleSize:  1,
-		PoolMaxSize:   poolmax,
-		CleanInterval: time.Second + time.Duration(c.triggers.Timers.Rand().Intn(1000))*time.Millisecond,
-		Backoff:       600*time.Millisecond + time.Duration(c.triggers.Timers.Rand().Intn(600))*time.Millisecond,
-		SecretKey:     c.Config.APIKey,
-		Handler:       remWs.Wrap(c.prefixURLbase(c.Config.Router), c.Logger.HTTPLog.Writer()).ServeHTTP,
-		Logger:        &tunnelLogger{Logger: c.Logger, sendSiteErrors: ci.User.DevAllowed},
-	})
-	c.Printf("Tunneling to %q with %d connections; cleaner:%s, backoff:%s, url: %s, hash: %s",
-		strings.Join(c.tunnel.Targets, ", "), c.tunnel.PoolMaxSize, c.tunnel.CleanInterval,
-		c.tunnel.Backoff, ci.User.TunnelURL, c.tunnel.GetID())
-	c.tunnel.Start(ctx)
-}
-
-// prefixURLbase adds a prefix to an http request.
-// We need this to fix websocket-tunneled requests
-// from the website when url base is not the default.
-func (c *Client) prefixURLbase(h http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if c.Config.URLBase == "" || c.Config.URLBase == "/" {
-			h.ServeHTTP(w, r)
-			return
-		}
-
-		r2 := new(http.Request)
-		*r2 = *r
-		r2.URL = new(url.URL)
-		*r2.URL = *r.URL
-		r2.URL.Path = path.Join(c.Config.URLBase, r.URL.Path)
-
-		if r.URL.RawPath != "" {
-			r2.URL.RawPath = path.Join(c.Config.URLBase, r.URL.RawPath)
-		}
-
-		h.ServeHTTP(w, r2)
-	})
-}
-
 // runWebServer starts the http or https listener.
 func (c *Client) runWebServer() {
 	defer c.CapturePanic()
@@ -184,7 +96,7 @@ func (c *Client) StopWebServer(ctx context.Context) error {
 	if c.tunnel != nil {
 		defer c.tunnel.Shutdown()
 	}
-
+	// Wait for any active requests before shutting down the tunnel.
 	if err := c.server.Shutdown(ctx); err != nil {
 		return fmt.Errorf("shutting down web server: %w", err)
 	}
@@ -237,31 +149,4 @@ func (r *responseWrapper) Hijack() (net.Conn, *bufio.ReadWriter, error) {
 func (n *netConnWrapper) Write(b []byte) (int, error) {
 	mnd.HTTPRequests.Add("Response Bytes", int64(len(b)))
 	return n.Conn.Write(b) //nolint:wrapcheck
-}
-
-// tunnelLogger lets us tune the logs from the mulery tunnel.
-type tunnelLogger struct {
-	mnd.Logger
-	// sendSiteErrors true sends tunnel errors to website as notifications.
-	sendSiteErrors bool
-}
-
-// Debugf prints a message with DEBUG prefixed.
-func (l *tunnelLogger) Debugf(format string, v ...interface{}) {
-	l.Logger.Debugf(format, v...)
-}
-
-// Errorf prints a message with ERROR prefixed.
-func (l *tunnelLogger) Errorf(format string, v ...interface{}) {
-	// this is why we dont just pass the interface in as-is.
-	if l.sendSiteErrors {
-		l.Logger.Errorf(format, v...)
-	} else {
-		l.Logger.ErrorfNoShare(format, v...)
-	}
-}
-
-// Printf prints a message with INFO prefixed.
-func (l *tunnelLogger) Printf(format string, v ...interface{}) {
-	l.Logger.Printf(format, v...)
 }
